+++ packages/video_player/video_player/test/video_player_initialization_test.dart
@@ -33,7 +33,7 @@ void main() {
 
     final controller = VideoPlayerController.networkUrl(
       Uri.parse('https://127.0.0.1'),
-      videoPlayerOptions: VideoPlayerOptions(webOptions: expected),
+      videoPlayerOptions: const VideoPlayerOptions(webOptions: expected),
     );
     await controller.initialize();
 
@@ -71,6 +71,26 @@ void main() {
       fakeVideoPlayerPlatform.viewTypes[controller.playerId],
       expected,
       reason: 'view type must be passed to the platform',
+    );
+  });
+
+  test('back buffer duration is forwarded to platform', () async {
+    const expectedBackBufferDurationMs = 20000;
+
+    final controller = VideoPlayerController.networkUrl(
+      Uri.parse('https://127.0.0.1'),
+      videoPlayerOptions: const VideoPlayerOptions(
+        backBufferDurationMs: expectedBackBufferDurationMs,
+      ),
+    );
+
+    await controller.initialize();
+
+    expect(
+      fakeVideoPlayerPlatform.videoPlayerOptions.last?.backBufferDurationMs,
+      expectedBackBufferDurationMs,
+      reason:
+          'backBufferDurationMs must be forwarded to the platform via VideoCreationOptions.videoPlayerOptions',
     );
   });
 }